+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Next
+- Add `Escape` shortcut for closing the window
+
 ## 1.0.2 - 2022 Nov 4
 - Fix tray icon button (and maybe Linux)